refacto(services): split api.js into productApi.js, userApi.js, authApi.js #30
